SparseAlgebra.jl Documentation

Functions

SparseAlgebra.invariant_subspace — Method

invariant_subspace(As; backend_algorithm)

Finds one common invariant subspace of the given matrices As using the provived default method for finding one invaiant subspace backend_algorithm.

SparseAlgebra.@sparse — Macro

@sparse(ex)

Converts an Array in ex to sparse representation

SparseAlgebra.func — Method

func(x)

Indentity function. Returns x.