Functions

Here the main interface of the program is listed.

SparseAlgebra.invariant_subspace — Function

invariant_subspace(As; backend_algorithm)

Finds one common invariant subspace of the given matrices As using the provived default method backend_algorithm. By default, invariant_subspace_1 is used.

Each matrix in As to be a subtype of AbstractSparseMatrix

SparseAlgebra.many_invariant_subspaces — Function

many_invariant_subspaces(As; backend_algorithm)

Finds several common invariant subspaces of the given matrices As using the provived default method backend_algorithm. By default, invariant_subspace_1 is used.

Each matrix in As to be a subtype of AbstractSparseMatrix

SparseAlgebra.find_basis — Function

find_basis(vectors; used_algorithm, initialprime)

Finds a basis of the algebra generated with the given vectors using the provived default method used_algorithm while reducing input coefficients modulo initialprime. By default, find_basis_1_β stands for the algorithm and 2147483647 for the initial reduction modulo.

Each element in vector to be a subtype of AbstractSparseObject