Structures

SparseAlgebra.Sparsik — Type

Sparsik{T<:Field}

Provides basic 1-dimensional sparse vector interface. Parametrized with the type of coefficient field T.

Supports these fast operations:

inner, scale, reduce, iterate.

One can construct Sparsik from a dense vector with the from_dense function.

SparseAlgebra.DOK_Sparsik — Type

DOK_Sparsik{T<:Field}

Provides basic 2-dimensional sparse vector interface, i.e Sparse Matrix. Parametrized with the type of coefficient field T.

Supports these fast operations:

inner, scale, reduce, iterate, matvec and matmat (Base.prod).

One can construct BidimSparsik from a dense matrix with the from_dense function.

SparseAlgebra.Subspacik — Type

Subspacik{T<:Field}

Provides basic linear subspace interface. Parametrized with the type of coefficient field T.

One can add a vector to the spanning vectors of a Subspacik instance via eat_sparsik! function or create a subspace from scratch with linear_span!