Types

Type system in this package goes as follows.

ExactODEReduction.AbstractSparseObject — Type

AbstractSparseObject{T}

Base interface for anything Sparse implemented in this project

T is the type of the ground field

ExactODEReduction.AbstractSparseVector — Type

AbstractSparseObject{T}

Base interface for 1-dimensional Sparse vector. An object which implements AbstractSparseVector can be treated as an element of Rⁿ vectorspace. Thus, some vectorspace operations and left matvec are supported for such objects.

T is the type of the ground field

ExactODEReduction.AbstractSparseMatrix — Type

AbstractSparseObject{T}

Base interface for 2-dimensional Sparse vector, i.e Sparse Matrix. An object which implements AbstractSparseMatrix can be treated as an element of Rⁿ⃰ⁿ vectorspace or as an element of a matrix algebra. Thus, some vectorspace operations and matrix prodducts are supported for such objects.

T is the type of the ground field