Interface

ExactODEReduction.find_reductions — Function

find_reductions(system::ODE; overQ=true, makepositive=false, loglevel=Logging.Info)

Finds reductions of the system corresponding to a Jordan-Hoelder filtration. This means that the reduction form a chain, and there are no extra intermediate reduction in this chain. In particular, if there exists at least one reduction, it will be found.

Arguments:

system is an ODE system given as ODE object,

overQ tells the algorithm to search for reductions over rational numbers,

makepositive tells the algorithm to search for reductions with positive coefficients.

To enable this argument, you should have Polymake.jl imported.

Example:

using ExactODEReduction
odes = @ODEsystem(
    x'(t) = x + y,
    y'(t) = x - y - z,
    z'(t) = 2x - z
)
find_reductions(odes)

ExactODEReduction.find_smallest_constrained_reduction — Function

find_smallest_constrained_reduction(system::ODE, observables; overQ=true, makepositive=false, loglevel=Logging.Info)

Finds the best linear reduction of the system. If there exists a reduction, it will be found. Arguments:

system is an ODE system given as ODE object,

observables is a list of linear functions of initial variables desired to be preserved by reduction,

overQ tells the algorithm to search for reductions over rational numbers,

makepositive tells the algorithm to search for reductions with positive coefficients;

To enable this argument, you should have Polymake.jl imported.

Example:

using ExactODEReduction
odes = @ODEsystem(
    x'(t) = x + y,
    y'(t) = x - y - z,
    z'(t) = 2x - z
)
find_smallest_constrained_reduction(odes, [y])

ExactODEReduction.find_some_reduction — Function

find_some_reduction(system::ODE, observables; overQ=true, makepositive=false, loglevel=Logging.Info)

Finds a nontrivial linear reduction of the system. If there exists a reduction, it will be found.

Arguments:

system is an ODE system given as ODE object,

overQ tells the algorithm to search for reductions over rational numbers,

makepositive tells the algorithm to search for reductions with positive coefficients;

To enable this argument, you should have Polymake.jl imported.

Example:

using ExactODEReduction
odes = @ODEsystem(
    x'(t) = x + y,
    y'(t) = x - y - z,
    z'(t) = 2x - z
)
find_some_reduction(odes)