Functions

Here the main interface of the program is listed.

ExactODEReduction.find_reductions — Function

find_reductions(system; backend_algorithm)

Finds reductions of the system corresponding to a Jordan-Hoelder filtration. This means that the reduction form a chain, and there are no extra intermediate reduction in this chain. In particular, if there exists at least one reduction, it will be found. Arguments:

system is a list of the right-hand sides of the system, the i-th element is the derivative of the i-th variable in the corresponding polynomial ring

backend_algorithm is the linear algebra algorithm used. By default, invariant_subspace_1.

Missing docstring.

Missing docstring for invariant_subspace. Check Documenter's build log for details.

Missing docstring.

Missing docstring for many_invariant_subspaces. Check Documenter's build log for details.

ExactODEReduction.find_basis — Function

find_basis(vectors; used_algorithm, initialprime)

Finds a basis of the algebra generated with the given vectors using the provived default method used_algorithm while reducing input coefficients modulo initialprime. By default, find_basis_1_β stands for the algorithm and 2147483647 for the initial reduction modulo.

Each element in vector to be a subtype of AbstractSparseObject

ExactODEReduction.find_radical_sup — Function

find_radical_sup(Algebra)

Finds a basis of the radical of given matrix Algebra.

ExactODEReduction.wiedemann_solve — Function

wiedemann_solve(A, b; proved)

Returns a solution x to linear system Ax = b under an assumtion of A being square and nonsingular using Wiedemann system-solving algorithm. By default, proved = true which calls 'squarenonsingulardeterministicwiedemann' otherwise calls 'squarenonsingularrandomizedwiedemann'.

ExactODEReduction.construct_jacobians — Function

construct_jacobians(system)

For the given system of polynomials in variables xi consturucts a set of matrices Aᵢ over number field such that the Jacobian J of the provided system can represented as the sum J = Aᵢxⁱ

ExactODEReduction.rational_reconstruction — Function

rational_reconstruction(a, m)

Rational number reconstruction implementation borrowed from CLUE and modified a bit to suit the 'Modern Computer Algebra' definitions. Returns a rational of QQ field in a canonical form that is congruent a modulo m

a, m are integers