Functions

Here the main interface of the program is listed.

SparseAlgebra.invariant_subspace — Function

invariant_subspace(As; backend_algorithm)

Finds one common invariant subspace of the given matrices As using the provived default method backend_algorithm. By default, invariant_subspace_1 is used.

Each matrix in As to be a subtype of AbstractSparseMatrix

SparseAlgebra.many_invariant_subspaces — Function

many_invariant_subspaces(As; backend_algorithm)

Finds several common invariant subspaces of the given matrices As using the provived default method backend_algorithm. By default, invariant_subspace_1 is used.

Each matrix in As to be a subtype of AbstractSparseMatrix

SparseAlgebra.find_basis — Function

find_basis(vectors; used_algorithm, initialprime)

Finds a basis of the algebra generated with the given vectors using the provived default method used_algorithm while reducing input coefficients modulo initialprime. By default, find_basis_1_β stands for the algorithm and 2147483647 for the initial reduction modulo.

Each element in vector to be a subtype of AbstractSparseObject

SparseAlgebra.find_radical_sup — Function

find_radical_sup(Algebra)

Finds a basis of the radical of given matrix Algebra.

SparseAlgebra.wiedemann_solve — Function

wiedemann_solve(A, b; proved)

Returns a solution x to linear system Ax = b under an assumtion of A being square and nonsingular using Wiedemann system-solving algorithm. By default, proved = true which calls 'squarenonsingulardeterministicwiedemann' otherwise calls 'squarenonsingularrandomizedwiedemann'.

SparseAlgebra.construct_jacobians — Function

construct_jacobians(system)

For the given system of polynomials in variables xi consturucts a set of matrices Aᵢ over number field such that the Jacobian J of the provided system can represented as the sum J = Aᵢxⁱ

SparseAlgebra.rational_reconstruction — Function

rational_reconstruction(a, m)

Rational number reconstruction implementation borrowed from CLUE and modified a bit to suit the 'Modern Computer Algebra' definitions. Returns a rational of QQ field in a canonical form that is congruent a modulo m

a, m are integers