Functions for finding reductions

ExactODEReduction.find_reductions — Function

find_reductions(system::ODE; overQ=true, makepositive=false, loglevel=Logging.Info)

Finds reductions of the system corresponding to a Jordan-Hoelder filtration. This means that the reduction form a chain, and there are no extra intermediate reduction in this chain. In particular, if there exists at least one reduction, it will be found.

Arguments:

system is an ODE system given as ODE object,

overQ tells the algorithm to search for reductions over rational numbers,

makepositive tells the algorithm to search for reductions with positive coefficients.

To enable this argument, you should have Polymake.jl imported.

Example:

julia> using ExactODEReduction
julia> odes = @ODEsystem(
    x'(t) = x + y,
    y'(t) = x - y - z,
    z'(t) = 2x - z
)
julia> find_reductions(odes)
2-element Vector{Dict{Symbol, Any}}:
 Dict(:new_system => y1'(t) = 0
, :new_vars => Nemo.fmpq_mpoly[y - z + x])
 Dict(:new_system => y2'(t) = -y1(t) - y2(t)
y1'(t) = -y1(t) - y2(t)
, :new_vars => Nemo.fmpq_mpoly[y, z - x])

ExactODEReduction.find_smallest_constrained_reduction — Function

find_smallest_constrained_reduction(system::ODE, observables; overQ=true, makepositive=false, loglevel=Logging.Info)

Finds the best linear reduction of the system. If there exists a reduction, it will be found. Arguments:

system is an ODE system given as ODE object,

observables is a list of linear functions of initial variables desired to be preserved by reduction,

overQ tells the algorithm to search for reductions over rational numbers,

makepositive tells the algorithm to search for reductions with positive coefficients;

To enable this argument, you should have Polymake.jl imported.

Example:

julia> using ExactODEReduction
julia> odes = @ODEsystem(
    x'(t) = x + y,
    y'(t) = x - y - z,
    z'(t) = 2x - z
)
julia> find_smallest_constrained_reduction(odes, [2x + z])
Dict{Symbol, Any} with 2 entries:
  :new_system => y2'(t) = 2*y1(t) - y2(t)…
  :new_vars   => Nemo.fmpq_mpoly[y + 3*x, z + 2*x]

ExactODEReduction.find_some_reduction — Function

find_some_reduction(system::ODE, observables; overQ=true, makepositive=false, loglevel=Logging.Info)

Finds a nontrivial linear reduction of the system. If there exists a reduction, it will be found.

Arguments:

system is an ODE system given as ODE object,

overQ tells the algorithm to search for reductions over rational numbers,

makepositive tells the algorithm to search for reductions with positive coefficients;

To enable this argument, you should have Polymake.jl imported.

Example:

julia> using ExactODEReduction
julia> odes = @ODEsystem(
    x'(t) = x + y,
    y'(t) = x - y - z,
    z'(t) = 2x - z
)
julia> find_some_reduction(odes)
Dict{Symbol, Any} with 2 entries:
  :new_system => y2'(t) = -y1(t) - y2(t)…
  :new_vars   => Nemo.fmpq_mpoly[y, z - x]

Functions for exploring found reductions

ExactODEReduction.new_system — Function

new_system(r::Reduction)

Returns the `ODE` object that defines the reduced system.

ExactODEReduction.new_vars — Function

new_system(r::Reduction)

Returns the dictionary of new macro-variables expressed
as linear combinations of the original variables.